requirements.txt
pandas==2.2.2
numpy==2.1.3
scikit-learn==1.6.1
joblib==1.4.2
python-dotenv==1.0.1
ollama==0.4.7
